==> RAG/3.Muti_Query_Vector_Retriever.ipynb <==
import json
import sys
import os

current_dir = os.getcwd()
parent_dir = os.path.dirname(current_dir)
sys.path.append(parent_dir)
from DataProcessing.utils import load_yaml
from DataProcessing.extract_graphstate import (
    extract_documents_for_multivectorstore,
    extract_documents_for_multidocstore,
)

config = load_yaml("../config/embedding.yaml")
category_id = config["settings"]["category_id"]
filetype = config["settings"]["filetype"]
edit_path = config["settings"]["edit_path"]
output_path = config["settings"]["output_path"]
os.makedirs(output_path, exist_ok=True)


def load_json_files(path):
    data_list = []
    for filename in os.listdir(path):
        if filename.endswith(".json"):
            with open(os.path.join(path, filename), "r", encoding="utf-8") as file:
                data = json.load(file)
                data_list.append(data)
    return data_list


data_list = []
for category in config["settings"]["category_id"]:
    category_path = os.path.join(edit_path, category, "json")
    data_list.extend(load_json_files(category_path))


vectorstore_documents = []
docstore_documents = []
for data in data_list:
    docstore_document = extract_documents_for_multidocstore(data)
    docstore_documents.extend(docstore_document)
    vectorstore_document = extract_documents_for_multivectorstore(data)
    vectorstore_documents.extend(vectorstore_document)

import uuid
from datetime import datetime
from langchain.storage import InMemoryStore
from langchain_chroma import Chroma
from langchain.retrievers.multi_vector import MultiVectorRetriever
from langchain.retrievers.multi_query import MultiQueryRetriever
from langchain_openai import ChatOpenAI
from langchain_upstage.embeddings import UpstageEmbeddings
from langchain.retrievers import EnsembleRetriever
from langchain.storage import LocalFileStore
from langchain.embeddings import CacheBackedEmbeddings
from dotenv import load_dotenv
from langchain_core.documents import Document
from langchain_community.vectorstores.utils import filter_complex_metadata

load_dotenv()
time = datetime.now().strftime("%Y.%m.%d")
passage_embeddings = UpstageEmbeddings(model="solar-embedding-1-large-passage")

proj_name = f"2024.09.14_for_multivector_store"
store = LocalFileStore(
    f"/Users/youngseoklee/Desktop/workplace/MacroAgent-withRAG/cache/{proj_name}/data"
)

cached_embedder = CacheBackedEmbeddings.from_bytes_store(
    underlying_embeddings=passage_embeddings,
    document_embedding_cache=store,
    namespace=passage_embeddings.model,
)
DB_PATH = f"/Users/youngseoklee/Desktop/workplace/MacroAgent-withRAG/db/2024.09.14_for_multivector_store"
db = Chroma(persist_directory=DB_PATH, embedding_function=cached_embedder)
llm = ChatOpenAI(model="gpt-4o-mini", temperature=0)
id_key = "id_key"
doc_ids = [str(uuid.uuid4()) for _ in docstore_documents]


multi_query_retriever = MultiQueryRetriever.from_llm(
    retriever=db.as_retriever(),
    llm=llm,
)

multi_vector_retriever = MultiVectorRetriever(
    vectorstore=db,
    byte_store=store,
    id_key=id_key,
)

ensemble_retriever = EnsembleRetriever(
    retrievers=[multi_query_retriever, multi_vector_retriever], weights=[0.5, 0.5]
)


summary_docs = [
    Document(page_content=s.page_content, metadata={id_key: doc_ids[i], **s.metadata})
    for i, s in enumerate(vectorstore_documents)
]
filtered_summary_docs = filter_complex_metadata(documents=summary_docs)
filtered_docstore_documents = filter_complex_metadata(documents=docstore_documents)
multi_vector_retriever.vectorstore.add_documents(
    summary_docs, ids=[doc.metadata[id_key] for doc in summary_docs]
)
multi_vector_retriever.docstore.mset(list(zip(doc_ids, docstore_documents)))

retrieved_docs = multi_vector_retriever.invoke("옐런의 계획이 어떻게 되니?")
retrieved_docs

from langchain.callbacks.manager import CallbackManagerForRetrieverRun


def combined_retriever(query):
    # CallbackManagerForRetrieverRun 인스턴스 생성
    run_manager = CallbackManagerForRetrieverRun.get_noop_manager()

    multi_queries = multi_query_retriever.generate_queries(
        question=query, run_manager=run_manager
    )
    all_docs = []
    for q in multi_queries:
        docs = multi_vector_retriever.invoke(q)
        all_docs.extend(docs)

    # 중복 제거를 위해 딕셔너리를 사용
    unique_docs = {
        (doc.page_content, frozenset(doc.metadata.items())): doc for doc in all_docs
    }.values()

    return sorted(
        unique_docs, key=lambda x: x.metadata.get("relevance_score", 0), reverse=True
    )


query = "옐런의 계획이 어떻게 되니?"
combined_results = combined_retriever(query)
ensemble_results = ensemble_retriever.get_relevant_documents(query)

from langchain_core.runnables import RunnablePassthrough
from langchain_core.output_parsers import StrOutputParser
from langchain_openai import ChatOpenAI
from langchain_core.prompts import PromptTemplate
import pandas as pd
import sys
import os
import json
from langchain_anthropic import ChatAnthropic

current_dir = os.getcwd()
parent_dir = os.path.dirname(current_dir)
sys.path.append(parent_dir)
from DataProcessing.utils import load_yaml


path = "./data/custom_testdataset.xlsx"
# testset = pd.read_csv(path)
testset = pd.read_excel(path)

questions = testset["question"].to_list()
ground_truth = testset["ground_truth"].to_list()

data = {"question": [], "answer": [], "contexts": [], "ground_truth": ground_truth}

prompt_template = load_yaml("../prompts/Retriever._prompt.yaml")["prompt"]
prompt = PromptTemplate.from_template(prompt_template)
#llm = ChatOpenAI(model="gpt-4o-mini", temperature=0.5)
llm = ChatAnthropic(model="claude-3-5-sonnet-20240620", temperature=0.5)
rag_chain = (
    {"context": multi_vector_retriever, "question": RunnablePassthrough()}
    | prompt
    | llm
    | StrOutputParser()
)


for query in questions:
    data["question"].append(query)
    data["answer"].append(rag_chain.invoke(query))
    data["contexts"].append(
        [doc.page_content for doc in multi_vector_retriever.invoke(query)]
    )

path = "./data/customtestset_sonnet_multivector.json"
with open(path, "w") as file:
    json.dump(data, file)
